Add 'ProjectList' component for project filtering
Introduced a new 'ProjectList' component in ProjectList.tsx file. This component accepts a list of projects and a selected category as props and displays only the projects belonging to the selected category. If state value is 'All', it displays all projects. The purpose of this component is to provide filter functionality to the projects based on their categories.