Import React explicitly in layout.tsx
React has been explicitly imported in layout.tsx. This change is necessary after the upgrade of React to version 17, in which JSX Transform no longer requires React to be in scope. Therefore, the code will work even without this import. But it's better to explicitly import React to ensure code can be seamlessly ported to other environments that might not have such configurations.